Error handling implemented. Unused codes removed.
