+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/data/repository/network/interfaces/ICharacterRepository.kt
@@ -3,5 +3,5 @@ package com.pinheiro.marvelhqs.data.repository.network.interfaces
 import com.pinheiro.marvelhqs.data.repository.network.response.CharacterResponse
 
 interface ICharacterRepository {
-suspend fun getCharacter(limit: String, offset: String): CharacterResponse
+suspend fun getCharacter(limit: String, offset: String):Result<CharacterResponse>
 }
